Big Level Design Push
The majority of the floor plan and lighting have been established. Need to finish the rear "dead end" path and connection to the right side. After that, lighting goes throughout to unify everything and then possibly time for decorative finishes (tables/chairs? suits of armor? something on the walls (paintings?))